requirements.txt
langchain-community
langchain_openai
langchain
python-dotenv
openai
streamlit
pinecone
tqdm
pypdf2
